Updated code for emulators
